chore: use standard lint settings and fix lints (#4066)
Description
---
- updates lints.toml to match [tari-meta](https://github.com/tari-project/meta/blob/main/lints.toml)
- fixes a number of clippy minor lints

Motivation and Context
---
Tari repo uses standard linter settings EDIT: there are too many cast clippys to investigate, so that is left for another PR

Larger fixes are in #4063 #4064 #4065 which should be merged before this PR.

How Has This Been Tested?
---
`cargo lints clippy` passes

+++ base_layer/core/src/mempool/mempool_storage.rs
@@ -279,9 +279,9 @@ impl MempoolStorage {
     pub fn stats(&self) -> StatsResponse {
         let weighting = self.get_transaction_weighting(0);
         StatsResponse {
-            total_txs: self.len(),
-            unconfirmed_txs: self.unconfirmed_pool.len(),
-            reorg_txs: self.reorg_pool.len(),
+            total_txs: self.len() as u64,
+            unconfirmed_txs: self.unconfirmed_pool.len() as u64,
+            reorg_txs: self.reorg_pool.len() as u64,
             total_weight: self.unconfirmed_pool.calculate_weight(&weighting),
         }
     }
